+++ app/Http/Controllers/Api/CompaniesController.php
@@ -6,6 +6,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
 use App\Models\Company;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -15,9 +16,11 @@ class CompaniesController extends Controller
      * Display a listing of the resource.
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny');
         $companies = Company::latest()->paginate(25);
 
         return $companies;
@@ -29,10 +32,11 @@ class CompaniesController extends Controller
      * @param Request $request
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function store(Request $request)
     {
-        
+        $this->authorize('create');
         $company = Company::create($request->all());
 
         return response()->json($company, 201);
@@ -41,12 +45,14 @@ class CompaniesController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function show($id)
     {
+        $this->authorize('view');
         $company = Company::findOrFail($id);
 
         return $company;
@@ -56,13 +62,14 @@ class CompaniesController extends Controller
      * Update the specified resource in storage.
      *
      * @param Request $request
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(Request $request, $id)
     {
-        
+        $this->authorize('edit');
         $company = Company::findOrFail($id);
         $company->update($request->all());
 
@@ -72,12 +79,14 @@ class CompaniesController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('delete');
         Company::destroy($id);
 
         return response()->json(null, 204);
